+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ next
 - Added `reason` input to deleting a user, saves as field in DeletedUser entry.
 - Updated docker files to support production build with build args.
 - Hide delete user button when viewing the currently logged in user.
+- Manage Roles pages only allow assignment of roles that are available on a given domain or site, more info shown if no roles are available.
 
 1.2.0
 -----